bug-fix-auth-controller: fix auth controllers
- Fit auth controllers
- Fix unit tests
