+++ SocialReview.API/Controllers/AuthController.cs
@@ -7,11 +7,11 @@ namespace SocialReview.API.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class CustomerAuthController : ControllerBase
+    public class AuthController : ControllerBase
     {
         private readonly IAuthService _authService;
 
-        public CustomerAuthController(IAuthService authService)
+        public AuthController(IAuthService authService)
         {
             _authService = authService;
         }
@@ -32,11 +32,26 @@ namespace SocialReview.API.Controllers
         }
 
         /// <summary>
+        /// Registers a new establishment using the provided request data.
+        /// </summary>
+        /// <param name="request">A EstablishmentRegisterDto object containing the registration data.</param>
+        /// <returns>A Establishment object or a BadRequest if the registration fails.</returns>
+        [HttpPost("establishment-register")]
+        public async Task<ActionResult<Establishment>> EstablishmentRegister(EstablishmentRegisterDto request)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var establishment = await _authService.RegisterAsync<Establishment, EstablishmentRegisterDto>(request);
+            return Ok(establishment);
+        }
+
+        /// <summary>
         /// Logs in a user with the provided email and password.
         /// </summary>
         /// <param name="request">The user login data transfer object containing the email and password.</param>
         /// <returns>An ActionResult containing a token for the authenticated user or a BadRequest if the model state is invalid.</returns>
-        [HttpPost("customer-login")]
+        [HttpPost("login")]
         public async Task<ActionResult<string>> CustomerLogin(UserLoginDto request)
         {
             if (!ModelState.IsValid)
